docs(#8): constants 폴더 recommendation 파일 통일
