--- src/constants/HeroList.ts
@@ -1,26 +1,0 @@
-export const heroList = [
-  {
-    id: 1,
-    name: '채영대',
-    src: '/assets/images/recommendation/profile.svg',
-    selected: false,
-  },
-  {
-    id: 2,
-    name: '민혜린',
-    src: '/assets/images/recommendation/profile.svg',
-    selected: false,
-  },
-  {
-    id: 3,
-    name: '박형준',
-    src: '/assets/images/recommendation/profile.svg',
-    selected: false,
-  },
-  {
-    id: 4,
-    name: '이재영',
-    src: '/assets/images/recommendation/profile.svg',
-    selected: false,
-  },
-];
