--- src/constants/ProfileList.ts
@@ -1,22 +1,0 @@
-export const profiles = [
-  {
-    id: 1,
-    src: '/assets/images/recommendation/profile.svg',
-    name: '김지원',
-  },
-  {
-    id: 2,
-    src: '/assets/images/recommendation/profile.svg',
-    name: '박지원',
-  },
-  {
-    id: 3,
-    src: '/assets/images/recommendation/profile.svg',
-    name: '이예원',
-  },
-  {
-    id: 4,
-    src: '/assets/images/recommendation/profile.svg',
-    name: '김지원',
-  },
-];
