--- src/constants/TendencyQuestion.ts
@@ -1,41 +1,0 @@
-export const question1 = [
-  {
-    id: 1,
-    title: '피드백 스타일은 어땠나요?',
-    answer1: '솔직한 편',
-    answer2: '돌려말하는 편',
-  },
-  {
-    id: 2,
-    title: '어떤 스타일의 팀을 추구하셨나요?',
-    answer1: '업무에 집중',
-    answer2: '친목과 업무를 통해',
-  },
-  {
-    id: 3,
-    title: '외향적, 내향적 어디에 더 가까우신가요?',
-    answer1: '외향적',
-    answer2: '내향적',
-  },
-];
-
-export const question2 = [
-  {
-    id: 1,
-    title: '작업 스타일은 어땠나요?',
-    answer1: '신중하게 작업',
-    answer2: '속도감 있게 작업',
-  },
-  {
-    id: 2,
-    title: '결과와 과정, 어디에 더 중점을 두셨나요?',
-    answer1: '결과주의',
-    answer2: '과정주의',
-  },
-  {
-    id: 3,
-    title: '추구하는 워라밸은 어땠나요?',
-    answer1: '워커홀릭',
-    answer2: '워라밸 중시',
-  },
-];
