+++ src/constants/recommendation.ts
@@ -1,0 +1,104 @@
+export const keywordList = [
+  '계획 마스터',
+  '불타는 열정왕',
+  '전공스킬 넘사벽',
+  '소통의 귀재',
+  '논리의 왕',
+  '분위기 메이커',
+  '꽉찬 리더십',
+  '경험가득 고인물',
+  '올라운더',
+  '빛의 작업속도',
+];
+
+export const heroList = [
+  {
+    id: 1,
+    name: '채영대',
+    src: '/assets/images/recommendation/profile.svg',
+    selected: false,
+  },
+  {
+    id: 2,
+    name: '민혜린',
+    src: '/assets/images/recommendation/profile.svg',
+    selected: false,
+  },
+  {
+    id: 3,
+    name: '박형준',
+    src: '/assets/images/recommendation/profile.svg',
+    selected: false,
+  },
+  {
+    id: 4,
+    name: '이재영',
+    src: '/assets/images/recommendation/profile.svg',
+    selected: false,
+  },
+];
+
+export const profiles = [
+  {
+    id: 1,
+    src: '/assets/images/recommendation/profile.svg',
+    name: '김지원',
+  },
+  {
+    id: 2,
+    src: '/assets/images/recommendation/profile.svg',
+    name: '박지원',
+  },
+  {
+    id: 3,
+    src: '/assets/images/recommendation/profile.svg',
+    name: '이예원',
+  },
+  {
+    id: 4,
+    src: '/assets/images/recommendation/profile.svg',
+    name: '김지원',
+  },
+];
+
+export const question1 = [
+  {
+    id: 1,
+    title: '피드백 스타일은 어땠나요?',
+    answer1: '솔직한 편',
+    answer2: '돌려말하는 편',
+  },
+  {
+    id: 2,
+    title: '어떤 스타일의 팀을 추구하셨나요?',
+    answer1: '업무에 집중',
+    answer2: '친목과 업무를 통해',
+  },
+  {
+    id: 3,
+    title: '외향적, 내향적 어디에 더 가까우신가요?',
+    answer1: '외향적',
+    answer2: '내향적',
+  },
+];
+
+export const question2 = [
+  {
+    id: 1,
+    title: '작업 스타일은 어땠나요?',
+    answer1: '신중하게 작업',
+    answer2: '속도감 있게 작업',
+  },
+  {
+    id: 2,
+    title: '결과와 과정, 어디에 더 중점을 두셨나요?',
+    answer1: '결과주의',
+    answer2: '과정주의',
+  },
+  {
+    id: 3,
+    title: '추구하는 워라밸은 어땠나요?',
+    answer1: '워커홀릭',
+    answer2: '워라밸 중시',
+  },
+];
